making service for category, product and home APIs.
